+++ Pod/SyneriseSDK.framework/Headers/SNRProductRemovedFromCartEvent.h
@@ -11,13 +11,13 @@
 NS_ASSUME_NONNULL_BEGIN
 
 /**
- * @class SNRRemovedProductFromCartEvent
+ * @class SNRProductRemovedFromCartEvent
  *
  * Represents a 'client removed product from cart' event.
  */
 
-NS_SWIFT_NAME(RemovedProductFromCartEvent)
-@interface SNRRemovedProductFromCartEvent : SNRCartEvent
+NS_SWIFT_NAME(ProductRemovedFromCartEvent)
+@interface SNRProductRemovedFromCartEvent : SNRCartEvent
 
 @end